refactor(notify): extract config manager
