+++ crates/notify/src/integration.rs
@@ -12,6 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use crate::config_manager::{NotifyConfigManager, notify_configuration_hint};
 use crate::notification_system_subscriber::NotificationSystemSubscriberView;
 use crate::notifier::TargetList;
 use crate::runtime_facade::NotifyRuntimeFacade;
@@ -24,12 +25,7 @@ use crate::{
     rules::{BucketNotificationConfig, ParseConfigError},
 };
 use hashbrown::HashMap;
-use rustfs_config::notify::{
-    DEFAULT_NOTIFY_TARGET_STREAM_CONCURRENCY, ENV_NOTIFY_TARGET_STREAM_CONCURRENCY, ENV_NOTIFY_WEBHOOK_ENABLE,
-    ENV_NOTIFY_WEBHOOK_ENDPOINT, NOTIFY_AMQP_SUB_SYS, NOTIFY_KAFKA_SUB_SYS, NOTIFY_MQTT_SUB_SYS, NOTIFY_MYSQL_SUB_SYS,
-    NOTIFY_NATS_SUB_SYS, NOTIFY_POSTGRES_SUB_SYS, NOTIFY_PULSAR_SUB_SYS, NOTIFY_REDIS_SUB_SYS, NOTIFY_WEBHOOK_SUB_SYS,
-};
-use rustfs_config::{ENV_NOTIFY_ENABLE, EVENT_DEFAULT_DIR};
+use rustfs_config::notify::{DEFAULT_NOTIFY_TARGET_STREAM_CONCURRENCY, ENV_NOTIFY_TARGET_STREAM_CONCURRENCY};
 use rustfs_ecstore::config::{Config, KVS};
 use rustfs_s3_common::EventName;
 use rustfs_targets::arn::TargetID;
@@ -42,36 +38,6 @@ use tokio::sync::{RwLock, Semaphore, broadcast};
 use tracing::{debug, info, warn};
 
 const MAX_RECENT_LIVE_EVENTS: usize = 1024;
-
-fn notify_configuration_hint() -> String {
-    let webhook_enable_primary = format!("{ENV_NOTIFY_WEBHOOK_ENABLE}_PRIMARY");
-    let webhook_endpoint_primary = format!("{ENV_NOTIFY_WEBHOOK_ENDPOINT}_PRIMARY");
-    format!(
-        "No notify targets configured. Check {ENV_NOTIFY_ENABLE}=true and instance-scoped target env vars (for example {webhook_enable_primary} + {webhook_endpoint_primary} for arn:rustfs:sqs::primary:webhook). If using default queue_dir, ensure {EVENT_DEFAULT_DIR} is writable."
-    )
-}
-
-fn subsystem_target_type(target_type: &str) -> &str {
-    match target_type {
-        NOTIFY_AMQP_SUB_SYS => "amqp",
-        NOTIFY_WEBHOOK_SUB_SYS => "webhook",
-        NOTIFY_KAFKA_SUB_SYS => "kafka",
-        NOTIFY_MQTT_SUB_SYS => "mqtt",
-        NOTIFY_MYSQL_SUB_SYS => "mysql",
-        NOTIFY_NATS_SUB_SYS => "nats",
-        NOTIFY_POSTGRES_SUB_SYS => "postgres",
-        NOTIFY_PULSAR_SUB_SYS => "pulsar",
-        NOTIFY_REDIS_SUB_SYS => "redis",
-        _ => target_type,
-    }
-}
-
-fn runtime_target_id_for_subsystem(target_type: &str, target_name: &str) -> TargetID {
-    TargetID {
-        id: target_name.to_lowercase(),
-        name: subsystem_target_type(target_type).to_string(),
-    }
-}
 
 #[derive(Clone)]
 pub struct LiveEventBatch {
@@ -258,6 +224,10 @@ impl NotificationSystem {
         )
     }
 
+    fn config_manager(&self) -> NotifyConfigManager {
+        NotifyConfigManager::new(self.config.clone(), self.registry.clone(), self.notifier.clone(), self.runtime_facade())
+    }
+
     /// Creates a new NotificationSystem
     pub fn new(config: Config) -> Self {
         let concurrency_limiter =
@@ -279,29 +249,7 @@ impl NotificationSystem {
 
     /// Initializes the notification system
     pub async fn init(&self) -> Result<(), NotificationError> {
-        info!("Initialize notification system...");
-
-        let config = {
-            let guard = self.config.read().await;
-            debug!(
-                subsystem_count = guard.0.len(),
-                "Initializing notification system with configuration summary"
-            );
-            guard.clone()
-        };
-
-        let targets: Vec<Box<dyn Target<Event> + Send + Sync>> = self.registry.create_targets_from_config(&config).await?;
-
-        info!("{} notification targets were created", targets.len());
-        if targets.is_empty() {
-            warn!("{}", notify_configuration_hint());
-        }
-
-        // Initialize targets and start event streams
-        let activation = self.runtime_facade().activate_targets_with_replay(targets).await;
-        self.runtime_facade().replace_targets(activation).await?;
-        info!("Notification system initialized");
-        Ok(())
+        self.config_manager().init().await
     }
 
     /// Gets a list of Targets for all currently active (initialized).
@@ -351,35 +299,6 @@ impl NotificationSystem {
         history.snapshot_since(after_sequence, limit.max(1))
     }
 
-    async fn update_config_and_reload<F>(&self, mut modifier: F) -> Result<(), NotificationError>
-    where
-        F: FnMut(&mut Config) -> bool, // The closure returns a boolean value indicating whether the configuration has been changed
-    {
-        let Some(store) = rustfs_ecstore::global::new_object_layer_fn() else {
-            return Err(NotificationError::StorageNotAvailable(
-                "Failed to save target configuration: server storage not initialized".to_string(),
-            ));
-        };
-
-        let mut new_config = rustfs_ecstore::config::com::read_config_without_migrate(store.clone())
-            .await
-            .map_err(|e| NotificationError::ReadConfig(e.to_string()))?;
-
-        if !modifier(&mut new_config) {
-            // If the closure indication has not changed, return in advance
-            info!("Configuration not changed, skipping save and reload.");
-            return Ok(());
-        }
-
-        // Save the modified configuration to storage
-        rustfs_ecstore::config::com::save_server_config(store, &new_config)
-            .await
-            .map_err(|e| NotificationError::SaveConfig(e.to_string()))?;
-
-        info!("Configuration updated. Reloading system...");
-        self.reload_config(new_config).await
-    }
-
     /// Accurately remove a Target and its related resources through TargetID.
     ///
     /// This process includes:
@@ -393,28 +312,7 @@ impl NotificationSystem {
     /// # return
     /// If successful, return `Ok(())`.
     pub async fn remove_target(&self, target_id: &TargetID, target_type: &str) -> Result<(), NotificationError> {
-        info!("Attempting to remove target: {}", target_id);
-
-        let ttype = target_type.to_lowercase();
-        let tname = target_id.id.to_lowercase();
-
-        self.update_config_and_reload(|config| {
-            let mut changed = false;
-            if let Some(targets_of_type) = config.0.get_mut(&ttype) {
-                if targets_of_type.remove(&tname).is_some() {
-                    info!("Removed target {} from configuration", target_id);
-                    changed = true;
-                }
-                if targets_of_type.is_empty() {
-                    config.0.remove(&ttype);
-                }
-            }
-            if !changed {
-                warn!("Target {} not found in configuration", target_id);
-            }
-            changed
-        })
-        .await
+        self.config_manager().remove_target(target_id, target_type).await
     }
 
     /// Set or update a Target configuration.
@@ -430,14 +328,7 @@ impl NotificationSystem {
     /// If the target configuration is successfully set, it returns Ok(()).
     /// If the target configuration is invalid, it returns Err(NotificationError::Configuration).
     pub async fn set_target_config(&self, target_type: &str, target_name: &str, kvs: KVS) -> Result<(), NotificationError> {
-        info!("Setting config for target {} of type {}", target_name, target_type);
-        let ttype = target_type.to_lowercase();
-        let tname = target_name.to_lowercase();
-        self.update_config_and_reload(|config| {
-            config.0.entry(ttype.clone()).or_default().insert(tname.clone(), kvs.clone());
-            true // The configuration is always modified
-        })
-        .await
+        self.config_manager().set_target_config(target_type, target_name, kvs).await
     }
 
     /// Removes all notification configurations for a bucket.
@@ -464,82 +355,12 @@ impl NotificationSystem {
     /// If the target configuration is successfully removed, it returns Ok(()).
     /// If the target configuration does not exist, it returns Ok(()) without making any changes.
     pub async fn remove_target_config(&self, target_type: &str, target_name: &str) -> Result<(), NotificationError> {
-        info!("Removing config for target {} of type {}", target_name, target_type);
-
-        let ttype = target_type.to_lowercase();
-        let tname = target_name.to_lowercase();
-        let target_id = runtime_target_id_for_subsystem(&ttype, &tname);
-
-        // Deletion is prohibited if bucket rules refer to it
-        if self.notifier.is_target_bound_to_any_bucket(&target_id).await {
-            return Err(NotificationError::Configuration(format!(
-                "Target is still bound to bucket rules and deletion is prohibited: type={} name={}",
-                ttype, tname
-            )));
-        }
-
-        let config_result = self
-            .update_config_and_reload(|config| {
-                let mut changed = false;
-                if let Some(targets) = config.0.get_mut(&ttype) {
-                    if targets.remove(&tname).is_some() {
-                        changed = true;
-                    }
-                    if targets.is_empty() {
-                        config.0.remove(&ttype);
-                    }
-                }
-                if !changed {
-                    info!("Target {} of type {} not found, no changes made.", target_name, target_type);
-                }
-                debug!(
-                    subsystem_count = config.0.len(),
-                    "Target config removal processed and configuration summary updated"
-                );
-                changed
-            })
-            .await;
-
-        if config_result.is_ok() {
-            // Remove from target list
-            let target_list = self.notifier.target_list();
-            let mut target_list_guard = target_list.write().await;
-            let _ = target_list_guard.remove_target_only(&target_id).await;
-        }
-
-        config_result
-    }
-
-    /// Update configuration
-    async fn update_config(&self, new_config: Config) {
-        let mut config = self.config.write().await;
-        *config = new_config;
+        self.config_manager().remove_target_config(target_type, target_name).await
     }
 
     /// Reloads the configuration
     pub async fn reload_config(&self, new_config: Config) -> Result<(), NotificationError> {
-        info!("Reload notification configuration starts");
-
-        // Update the config
-        self.update_config(new_config.clone()).await;
-
-        // Create new targets from configuration
-        let targets: Vec<Box<dyn Target<Event> + Send + Sync>> = self
-            .registry
-            .create_targets_from_config(&new_config)
-            .await
-            .map_err(NotificationError::Target)?;
-
-        info!("{} notification targets were created from the new configuration", targets.len());
-        if targets.is_empty() {
-            warn!("{}", notify_configuration_hint());
-        }
-
-        // Initialize targets and start event streams using shared helper
-        let activation = self.runtime_facade().activate_targets_with_replay(targets).await;
-        self.runtime_facade().replace_targets(activation).await?;
-        info!("Configuration reloaded end");
-        Ok(())
+        self.config_manager().reload_config(new_config).await
     }
 
     /// Loads the bucket notification configuration
@@ -671,60 +492,5 @@ mod tests {
         assert!(batch.truncated);
         assert_eq!(batch.events.len(), 1);
         assert_eq!(batch.events[0].s3.object.key, "one");
-    }
-
-    #[test]
-    fn runtime_target_id_for_subsystem_maps_notify_webhook_to_runtime_type() {
-        let target_id = runtime_target_id_for_subsystem(NOTIFY_WEBHOOK_SUB_SYS, "Primary");
-        assert_eq!(target_id.id, "primary");
-        assert_eq!(target_id.name, "webhook");
-    }
-
-    #[test]
-    fn runtime_target_id_for_subsystem_maps_notify_amqp_to_runtime_type() {
-        let target_id = runtime_target_id_for_subsystem(NOTIFY_AMQP_SUB_SYS, "Primary");
-        assert_eq!(target_id.id, "primary");
-        assert_eq!(target_id.name, "amqp");
-    }
-
-    #[test]
-    fn runtime_target_id_for_subsystem_maps_notify_mqtt_to_runtime_type() {
-        let target_id = runtime_target_id_for_subsystem(NOTIFY_MQTT_SUB_SYS, "Analytics");
-        assert_eq!(target_id.id, "analytics");
-        assert_eq!(target_id.name, "mqtt");
-    }
-
-    #[test]
-    fn runtime_target_id_for_subsystem_maps_notify_kafka_to_runtime_type() {
-        let target_id = runtime_target_id_for_subsystem(NOTIFY_KAFKA_SUB_SYS, "EventBus");
-        assert_eq!(target_id.id, "eventbus");
-        assert_eq!(target_id.name, "kafka");
-    }
-
-    #[test]
-    fn runtime_target_id_for_subsystem_maps_notify_nats_to_runtime_type() {
-        let target_id = runtime_target_id_for_subsystem(NOTIFY_NATS_SUB_SYS, "Bus");
-        assert_eq!(target_id.id, "bus");
-        assert_eq!(target_id.name, "nats");
-    }
-
-    #[test]
-    fn runtime_target_id_for_subsystem_maps_notify_pulsar_to_runtime_type() {
-        let target_id = runtime_target_id_for_subsystem(NOTIFY_PULSAR_SUB_SYS, "Ledger");
-        assert_eq!(target_id.id, "ledger");
-        assert_eq!(target_id.name, "pulsar");
-    }
-
-    #[test]
-    fn runtime_target_id_for_subsystem_maps_notify_redis_to_runtime_type() {
-        let target_id = runtime_target_id_for_subsystem(NOTIFY_REDIS_SUB_SYS, "Primary");
-        assert_eq!(target_id.id, "primary");
-        assert_eq!(target_id.name, "redis");
-    }
-    #[test]
-    fn runtime_target_id_for_subsystem_maps_notify_postgres_to_runtime_type() {
-        let target_id = runtime_target_id_for_subsystem(NOTIFY_POSTGRES_SUB_SYS, "AuditTrail");
-        assert_eq!(target_id.id, "audittrail");
-        assert_eq!(target_id.name, "postgres");
     }
 }
